requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
strawberry-graphql==0.217.1
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.5
pydantic==2.5.1
